+++ engine/Core/src/core/xml/XMLDocument.cpp
@@ -7,7 +7,7 @@
 #pragma comment(lib, "xmllite.lib")
 
 namespace Ghurund::Core {
-	void XMLDocument::parse(const void* data, uint32_t size) {
+	void XMLDocument::parse(const void* data, size_t size) {
 		root.set(nullptr);
 		CreateXmlReader(__uuidof(IXmlReader), (void**)&reader, nullptr);
 		Microsoft::WRL::ComPtr<IStream> memStream = SHCreateMemStream((const BYTE*)data, (UINT)size);
